Move jit_build_settings.hpp and jit_build_options.hpp out of api/
jit_build_settings.hpp is a header that allows the kernel build to pass options
to the jit compiler, while jit_build_options.hpp contains some internal
configuration for the JIT build. Neither should be exposed through the API.

Move them into tt_metal/jit_build, alongside jit_build_options.cpp.
