+++ tt_metal/api/tt-metalium/kernel.hpp
@@ -23,8 +23,6 @@
 #include <tt-metalium/base_types.hpp>
 #include <tt-metalium/core_coord.hpp>
 #include <tt-metalium/hal_types.hpp>
-#include <tt-metalium/jit_build_options.hpp>
-#include <tt-metalium/jit_build_settings.hpp>
 #include <tt-metalium/kernel_types.hpp>
 #include <tt-metalium/runtime_args_data.hpp>
 #include <tt-metalium/tt_backend_api_types.hpp>
@@ -43,7 +41,6 @@ namespace tt {
 namespace tt_metal {
 
 class IDevice;
-class JitBuildOptions;
 enum class DataMovementProcessor;
 class KernelImpl;
