+++ tt_metal/impl/program/program.cpp
@@ -51,7 +51,7 @@
 #include "hal.hpp"
 #include "hal_types.hpp"
 #include "jit_build/build.hpp"
-#include "jit_build_options.hpp"
+#include "jit_build/jit_build_options.hpp"
 #include "kernel.hpp"
 #include "kernel_types.hpp"
 #include "lightmetal/host_api_capture_helpers.hpp"
